+++ QuizApp/Scenes/Home/ViewModel/HomeViewModel.swift
@@ -10,18 +10,7 @@ import Foundation
 class HomeViewModel {
     
     var subjects: [Subject] = []
-    
-    func readJSONFile() {
-        if let url = Bundle.main.url(forResource: "quizzes", withExtension: "json") {
-            do {
-                let data = try Data(contentsOf: url)
-                let decoder = JSONDecoder()
-                self.subjects = try decoder.decode([Subject].self, from: data)
-            } catch {
-                print("Error decoding JSON: \(error)")
-            }
-        } else {
-            print("JSON file not found")
-        }
+    func loadJSONFile() {
+        self.subjects = JSONLoader.loadSubjectsFromJSONFile()
     }
 }
